chore: update sql schema definition

+++ backend/src/main/java/com/davnig/invoicemanagementapi/model/entity/FlatRateScheme.java
@@ -16,6 +16,6 @@ public class FlatRateScheme {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private int id;
     private String name;
-    private int rate;
+    private float rate;
 
 }
